backend/utils.py
import os
from dotenv import load_dotenv

load_dotenv()

ALLOWED_ORIGINS = [o.strip() for o in os.getenv("ALLOWED_ORIGINS", "*").split(",") if o.strip()]
MOCK_OPENAI = os.getenv("MOCK_OPENAI", "true").lower() == "true"
MOCK_ALPACA = os.getenv("MOCK_ALPACA", "true").lower() == "true"
OPENAI_API_KEY = os.getenv("OPENAI_API_KEY")
ALPACA_API_KEY = os.getenv("ALPACA_API_KEY")
ALPACA_SECRET_KEY = os.getenv("ALPACA_SECRET_KEY")
ALPACA_BASE_URL = os.getenv("ALPACA_BASE_URL", "https://paper-api.alpaca.markets")
